# api-docs.yaml
openapi: 3.1.0
info:
  title: itestra Dashboard API
  version: v0
servers:
- url: http://localhost:9000
  description: Generated server url
tags:
- name: Parking
  description: Parkingspot reservation modul
paths:
  /entries:
    get:
      tags:
      - Parking
      summary: Retrieve reserved parking sponts
      description: Returns all reserved parking sponts
      operationId: getEntries
      responses:
        "200":
          description: List of reserved parking spots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingEntry"
    post:
      tags:
      - Parking
      operationId: createEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParkingEntryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ParkingEntry"
  /overview:
    get:
      tags:
      - Parking
      operationId: getOverview
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingEntriesForDayResponse"
  /entries/{id}:
    delete:
      tags:
      - Parking
      operationId: deleteEntry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    ParkingEntryRequest:
      type: object
      description: DTO to request a parking spot
      properties:
        employeeName:
          type: string
          description: Name of the employee
          example: Max Mustermann
          minLength: 1
        date:
          type: string
          format: date
          description: Date formated with dd-MM-yyyy
          example: 24.12.2025
      required:
      - date
      - employeeName
    ParkingEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier
          example: 2
        employeeName:
          type: string
          description: Name of the employee
          example: Max Mustermann
        Reservation date:
          type: string
          format: date
          example: 24.12.2025
    ParkingEntriesForDayResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        entries:
          type: array
          items:
            $ref: "#/components/schemas/ParkingEntry"